# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type PricingType = typeof PricingType[keyof typeof PricingType];


export const PricingType = {
  Free: 'Free',
  Freemium: 'Freemium',
  Paid: 'Paid',
} as const;

export interface Category {
  id: number;
  slug: string;
  name: string;
  /** Lucide icon name representing the category */
  icon: string;
  description?: string;
  toolCount: number;
}

export interface Tool {
  id: number;
  slug: string;
  name: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  logoUrl: string;
  websiteUrl: string;
  categorySlug: string;
  categoryName: string;
  pricing: PricingType;
  featured: boolean;
  rating: number;
  tags: string[];
  /** @nullable */
  launchYear?: number | null;
}

export interface ToolListResult {
  items: Tool[];
  total: number;
  page: number;
  limit: number;
}

export interface DirectoryStats {
  totalTools: number;
  totalCategories: number;
  freeTools: number;
  freemiumTools: number;
  paidTools: number;
}

export type ListToolsParams = {
/**
 * Full-text search across tool name and description
 */
search?: string;
/**
 * Filter by category slug
 */
category?: string;
/**
 * Filter by pricing type
 */
pricing?: PricingType;
sort?: ListToolsSort;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListToolsSort = typeof ListToolsSort[keyof typeof ListToolsSort];


export const ListToolsSort = {
  popular: 'popular',
  newest: 'newest',
  name: 'name',
} as const;