# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: categories
    description: AI tool categories
  - name: tools
    description: AI tools directory
  - name: stats
    description: Directory-wide statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      description: Returns every category with a live count of tools in it
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /tools:
    get:
      operationId: listTools
      tags: [tools]
      summary: List AI tools
      description: Search and filter AI tools with pagination, ready to scale to thousands of entries
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Full-text search across tool name and description
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: Filter by category slug
        - name: pricing
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PricingType"
          description: Filter by pricing type
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [popular, newest, name]
            default: popular
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 24
      responses:
        "200":
          description: Paginated list of tools
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolListResult"
  /tools/featured:
    get:
      operationId: listFeaturedTools
      tags: [tools]
      summary: Featured tools
      description: Small curated set of featured/trending tools for homepage highlights
      responses:
        "200":
          description: Featured tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tool"
  /tools/{slug}:
    get:
      operationId: getTool
      tags: [tools]
      summary: Get a single tool by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tool detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stats:
    get:
      operationId: getDirectoryStats
      tags: [stats]
      summary: Directory-wide statistics
      description: Aggregate counts used for homepage/marketing stats
      responses:
        "200":
          description: Directory stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectoryStats"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    PricingType:
      type: string
      enum: [Free, Freemium, Paid]
    Category:
      type: object
      required: [id, slug, name, icon, toolCount]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        icon:
          type: string
          description: Lucide icon name representing the category
        description:
          type: string
        toolCount:
          type: integer
    Tool:
      type: object
      required:
        - id
        - slug
        - name
        - description
        - logoUrl
        - websiteUrl
        - categorySlug
        - categoryName
        - pricing
        - featured
        - rating
        - tags
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        longDescription:
          type: ["string", "null"]
        logoUrl:
          type: string
        websiteUrl:
          type: string
        categorySlug:
          type: string
        categoryName:
          type: string
        pricing:
          $ref: "#/components/schemas/PricingType"
        featured:
          type: boolean
        rating:
          type: number
        tags:
          type: array
          items:
            type: string
        launchYear:
          type: ["integer", "null"]
    ToolListResult:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Tool"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    DirectoryStats:
      type: object
      required: [totalTools, totalCategories, freeTools, freemiumTools, paidTools]
      properties:
        totalTools:
          type: integer
        totalCategories:
          type: integer
        freeTools:
          type: integer
        freemiumTools:
          type: integer
        paidTools:
          type: integer