# lib/api-zod/src/generated/types/tool.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PricingType } from './pricingType';

export interface Tool {
  id: number;
  slug: string;
  name: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  logoUrl: string;
  websiteUrl: string;
  categorySlug: string;
  categoryName: string;
  pricing: PricingType;
  featured: boolean;
  rating: number;
  tags: string[];
  /** @nullable */
  launchYear?: number | null;
}